COMMIT FIX
An accidental clicking of the revert button destroyed the last repository and it's history, and I am not yet well enough equipped with github to undo that. So I proceeded to make it worse.

This doesn't have the history, but generally:

Adds gui that changes screen
Adds aspect-ratio-safe scaling of game objects
Adds drawing of the main player on the game screen