Rescaleredo (#4)
* DRAW PLAYER

Adds the ability to draw the player properly with proper JComponent bounds and sprite to the center of the game screen.

* RESIZE-SAFE ENTITY POSITION

Adds resize-safe entity positioning that uses a scale value to place the entity within the content pane. Still cannot figure out how to make sure that this is enforced when resizing and different ratios.

* RESIZING AND RESCALING FIXES

Adds rescaling that does not particularly work. Last commit of branch before new rescaling logic replacing this old stuff.

* BETTER RESCALE

Adds new system for rescaling with logical coordinates and ends up with the same effects as before with less, better, and more organized code.

* MOVEMENT WITH RESIZING

Adds movement for the player on the game screen that is resized correctly and controlled by keyboard inputs using input map and action map.

* SMALL FIX

Small fix to stop the timer when button is pressed to allow for new input when already moving. Some movement is still janky and needs to be fixed.

* CODE REORG

Small reorganization.