MENU AND COMPONENTS
Adds main menu screen to the j frame for testing.
Adds component builder that can build Swing components of all kinds. Neat!
Adds better GUIConstants to be used by the builder.
Adds the start button to the main menu with a brief output for testing.

Removes the button builder class in favor of the more generic Component Builder (experimenting with generics :D)